Moves S3ClientConfig into s3 package.
