+++ apps/bfd-pipeline/bfd-pipeline-shared-utils/src/main/java/gov/cms/bfd/pipeline/sharedutils/s3/AwsS3ClientFactory.java
@@ -1,6 +1,5 @@
 package gov.cms.bfd.pipeline.sharedutils.s3;
 
-import gov.cms.bfd.pipeline.sharedutils.S3ClientConfig;
 import gov.cms.bfd.sharedutils.config.AwsClientConfig;
 import lombok.AllArgsConstructor;
 import software.amazon.awssdk.awscore.defaultsmode.DefaultsMode;
